Rename file to match module name
