+++ lib/manageiq/api/common.rb
@@ -1,8 +1,8 @@
-require "manageiq/api/common/action_dispatch-routing-option_redirect"
 require "manageiq/api/common/engine"
 require "manageiq/api/common/inflections"
 require "manageiq/api/common/logging"
 require "manageiq/api/common/metrics"
+require "manageiq/api/common/option_redirect_enhancements"
 require "manageiq/api/common/request"
 require "manageiq/api/common/routing"
 require "manageiq/api/common/user"
