+++ lib/manageiq/api/common/option_redirect_enhancements.rb
